int param changed to long

+++ CustomAdapterDemo/app/src/main/java/com/upsin/customadapterdemo/modelo/Persistence.java
@@ -8,6 +8,6 @@ public interface Persistence {
   public void closeDataBase();
   public long insertStudent(Student student);
   public long updateStudent(Student student);
-  public void deleteStudent(int id);
+  public void deleteStudent(long id);
 
 }
